splitting module in to many
